requirements.txt
pytz
httpx[http2]
python-dotenv
python-telegram-bot
python-telegram-bot-pagination